editoast: move impl Cache for Electrification to infra_cache
