+++ editoast/src/infra_cache/object_cache.rs
@@ -1,5 +1,6 @@
 mod buffer_stop_cache;
 mod detector_cache;
+mod electrification_cache;
 
 pub use buffer_stop_cache::BufferStopCache;
 pub use detector_cache::DetectorCache;
