+++ editoast/src/infra_cache/object_cache/electrification_cache.rs
@@ -1,0 +1,13 @@
+use crate::infra_cache::Cache;
+use crate::infra_cache::ObjectCache;
+use crate::schema::Electrification;
+
+impl Cache for Electrification {
+    fn get_track_referenced_id(&self) -> Vec<&String> {
+        self.track_ranges.iter().map(|tr| &*tr.track).collect()
+    }
+
+    fn get_object_cache(&self) -> ObjectCache {
+        ObjectCache::Electrification(self.clone())
+    }
+}
